Added SmartString
