+++ course2/meeting4/Makefile
@@ -1,2 +1,2 @@
-string_test: string_test.cc
+string_test: string_test.cc SmartString.h
 	g++ -o string_test -g -Wall -std=c++11 string_test.cc
